Create README.md
Rainfall Prediction is a complex and uncertain task that significantly impacts human society. 

Well-timed and accurate forecasting can proactively help reduce human and financial loss. 

This project presents a set of experiments that use standard machine learning methods to build models that can predict whether it will rain tomorrow or not based on the weather data for that day in major cities in Australia.